Create negative-results.csl
We are a new scientific journal, Negative Results. Negative findings are still a low priority for publication, so we need to find ways to make publishing them more attractive. Negative Results is a rigorously peer-reviewed edited, open access journal publishing negative findings of the highest quality with dubious or paradigm-shifting impact.